restructuring and alembic basic init
