--- test_app/alembic/versions/2022_08_06_1902-db0ae06097d4_handwritten_migration.py
@@ -1,25 +1,0 @@
-"""Handwritten migration
-
-Revision ID: db0ae06097d4
-Revises: e710ba02adb5
-Create Date: 2022-08-06 19:02:38.403329
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'db0ae06097d4'
-down_revision = 'e710ba02adb5'
-branch_labels = None
-depends_on = None
-
-# this is a handwritten migration 
-
-def upgrade() -> None:
-    pass
-
-
-def downgrade() -> None:
-    pass
