--- test_app/alembic/versions/383dd5c35290_test_revision_exclusion_of_tables_and_.py
@@ -1,28 +1,0 @@
-"""test revision exclusion of tables and delete column
-
-Revision ID: 383dd5c35290
-Revises: 38ecca09181c
-Create Date: 2022-08-05 00:44:43.319532
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision = '383dd5c35290'
-down_revision = '38ecca09181c'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'test_alembic')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('test_alembic', mysql.INTEGER(), autoincrement=False, nullable=True))
-    # ### end Alembic commands ###
