--- test_app/alembic/versions/38ecca09181c_test_upgrade_revision.py
@@ -1,28 +1,0 @@
-"""test upgrade revision
-
-Revision ID: 38ecca09181c
-Revises: cd330f20026d
-Create Date: 2022-08-04 17:53:24.229203
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '38ecca09181c'
-down_revision = 'cd330f20026d'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('test_alembic', sa.Integer(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'test_alembic')
-    # ### end Alembic commands ###
