--- test_app/alembic/versions/cd330f20026d_initial_migration.py
@@ -1,299 +1,0 @@
-"""Initial  migration
-
-Revision ID: cd330f20026d
-Revises: 
-Create Date: 2022-08-04 17:33:27.066900
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'cd330f20026d'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('admins',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_admins_id'), 'admins', ['id'], unique=False)
-    op.create_index(op.f('ix_admins_name'), 'admins', ['name'], unique=True)
-    op.create_table('class_types',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_class_types_id'), 'class_types', ['id'], unique=False)
-    op.create_index(op.f('ix_class_types_name'), 'class_types', ['name'], unique=False)
-    op.create_table('groups',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('number', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id', 'number')
-    )
-    op.create_index(op.f('ix_groups_id'), 'groups', ['id'], unique=False)
-    op.create_index(op.f('ix_groups_number'), 'groups', ['number'], unique=False)
-    op.create_table('groups_busy',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('group_id', sa.Integer(), nullable=False),
-    sa.Column('weekday', sa.String(length=255), nullable=False),
-    sa.Column('lesson', sa.Integer(), nullable=False),
-    sa.Column('is_busy', sa.Boolean(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_groups_busy_group_id'), 'groups_busy', ['group_id'], unique=False)
-    op.create_index(op.f('ix_groups_busy_id'), 'groups_busy', ['id'], unique=False)
-    op.create_index(op.f('ix_groups_busy_is_busy'), 'groups_busy', ['is_busy'], unique=False)
-    op.create_index(op.f('ix_groups_busy_lesson'), 'groups_busy', ['lesson'], unique=False)
-    op.create_index(op.f('ix_groups_busy_weekday'), 'groups_busy', ['weekday'], unique=False)
-    op.create_table('lessons',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('number', sa.Integer(), nullable=False),
-    sa.Column('time', sa.String(length=255), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_lessons_id'), 'lessons', ['id'], unique=False)
-    op.create_index(op.f('ix_lessons_number'), 'lessons', ['number'], unique=True)
-    op.create_index(op.f('ix_lessons_time'), 'lessons', ['time'], unique=True)
-    op.create_table('modules',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('year', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_modules_id'), 'modules', ['id'], unique=False)
-    op.create_index(op.f('ix_modules_name'), 'modules', ['name'], unique=False)
-    op.create_index(op.f('ix_modules_year'), 'modules', ['year'], unique=False)
-    op.create_table('rooms_busy',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('room_id', sa.Integer(), nullable=False),
-    sa.Column('weekday', sa.String(length=255), nullable=False),
-    sa.Column('lesson', sa.Integer(), nullable=False),
-    sa.Column('is_busy', sa.Boolean(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_rooms_busy_id'), 'rooms_busy', ['id'], unique=False)
-    op.create_index(op.f('ix_rooms_busy_is_busy'), 'rooms_busy', ['is_busy'], unique=False)
-    op.create_index(op.f('ix_rooms_busy_lesson'), 'rooms_busy', ['lesson'], unique=False)
-    op.create_index(op.f('ix_rooms_busy_room_id'), 'rooms_busy', ['room_id'], unique=False)
-    op.create_index(op.f('ix_rooms_busy_weekday'), 'rooms_busy', ['weekday'], unique=False)
-    op.create_table('teachers',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('last_name', sa.String(length=255), nullable=False),
-    sa.Column('first_name', sa.String(length=255), nullable=False),
-    sa.Column('second_name', sa.String(length=255), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_teachers_first_name'), 'teachers', ['first_name'], unique=False)
-    op.create_index(op.f('ix_teachers_id'), 'teachers', ['id'], unique=False)
-    op.create_index(op.f('ix_teachers_last_name'), 'teachers', ['last_name'], unique=False)
-    op.create_index(op.f('ix_teachers_second_name'), 'teachers', ['second_name'], unique=False)
-    op.create_table('teachers_busy',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('teacher_id', sa.Integer(), nullable=False),
-    sa.Column('weekday', sa.String(length=255), nullable=False),
-    sa.Column('lesson', sa.Integer(), nullable=False),
-    sa.Column('is_busy', sa.Boolean(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_teachers_busy_id'), 'teachers_busy', ['id'], unique=False)
-    op.create_index(op.f('ix_teachers_busy_is_busy'), 'teachers_busy', ['is_busy'], unique=False)
-    op.create_index(op.f('ix_teachers_busy_lesson'), 'teachers_busy', ['lesson'], unique=False)
-    op.create_index(op.f('ix_teachers_busy_teacher_id'), 'teachers_busy', ['teacher_id'], unique=False)
-    op.create_index(op.f('ix_teachers_busy_weekday'), 'teachers_busy', ['weekday'], unique=False)
-    op.create_table('weekdays',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_weekdays_id'), 'weekdays', ['id'], unique=False)
-    op.create_index(op.f('ix_weekdays_name'), 'weekdays', ['name'], unique=True)
-    op.create_table('years',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('number', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id', 'number')
-    )
-    op.create_index(op.f('ix_years_id'), 'years', ['id'], unique=False)
-    op.create_index(op.f('ix_years_number'), 'years', ['number'], unique=False)
-    op.create_table('modules_typesOfClasses_association',
-    sa.Column('Module_id', sa.Integer(), nullable=True),
-    sa.Column('TypeOfClass_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['Module_id'], ['modules.id'], ),
-    sa.ForeignKeyConstraint(['TypeOfClass_id'], ['class_types.id'], )
-    )
-    op.create_table('rooms',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('number', sa.Integer(), nullable=False),
-    sa.Column('class_type_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['class_type_id'], ['class_types.id'], ),
-    sa.PrimaryKeyConstraint('id', 'number')
-    )
-    op.create_index(op.f('ix_rooms_class_type_id'), 'rooms', ['class_type_id'], unique=False)
-    op.create_index(op.f('ix_rooms_id'), 'rooms', ['id'], unique=False)
-    op.create_index(op.f('ix_rooms_number'), 'rooms', ['number'], unique=False)
-    op.create_table('students',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('last_name', sa.String(length=255), nullable=False),
-    sa.Column('first_name', sa.String(length=255), nullable=False),
-    sa.Column('second_name', sa.String(length=255), nullable=True),
-    sa.Column('academic_year', sa.Integer(), nullable=True),
-    sa.Column('academic_group', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['academic_group'], ['groups.number'], ),
-    sa.ForeignKeyConstraint(['academic_year'], ['years.number'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_students_academic_group'), 'students', ['academic_group'], unique=False)
-    op.create_index(op.f('ix_students_academic_year'), 'students', ['academic_year'], unique=False)
-    op.create_index(op.f('ix_students_first_name'), 'students', ['first_name'], unique=False)
-    op.create_index(op.f('ix_students_id'), 'students', ['id'], unique=False)
-    op.create_index(op.f('ix_students_last_name'), 'students', ['last_name'], unique=False)
-    op.create_index(op.f('ix_students_second_name'), 'students', ['second_name'], unique=False)
-    op.create_table('teachers_modules_association',
-    sa.Column('Teacher_id', sa.Integer(), nullable=True),
-    sa.Column('Module_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['Module_id'], ['modules.id'], ),
-    sa.ForeignKeyConstraint(['Teacher_id'], ['teachers.id'], )
-    )
-    op.create_table('weekdays_lessons_association',
-    sa.Column('Weekday_id', sa.Integer(), nullable=True),
-    sa.Column('Lesson_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['Lesson_id'], ['lessons.id'], ),
-    sa.ForeignKeyConstraint(['Weekday_id'], ['weekdays.id'], )
-    )
-    op.create_table('schedule',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('semester', sa.Integer(), nullable=True),
-    sa.Column('group', sa.Integer(), nullable=True),
-    sa.Column('weekday', sa.String(length=255), nullable=False),
-    sa.Column('lesson_number', sa.Integer(), nullable=False),
-    sa.Column('module_id', sa.Integer(), nullable=True),
-    sa.Column('class_type', sa.String(length=255), nullable=True),
-    sa.Column('room', sa.Integer(), nullable=True),
-    sa.Column('teacher_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['class_type'], ['class_types.name'], ),
-    sa.ForeignKeyConstraint(['group'], ['groups.number'], ),
-    sa.ForeignKeyConstraint(['lesson_number'], ['lessons.number'], ),
-    sa.ForeignKeyConstraint(['module_id'], ['modules.id'], ),
-    sa.ForeignKeyConstraint(['room'], ['rooms.number'], ),
-    sa.ForeignKeyConstraint(['teacher_id'], ['teachers.id'], ),
-    sa.ForeignKeyConstraint(['weekday'], ['weekdays.name'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_schedule_class_type'), 'schedule', ['class_type'], unique=False)
-    op.create_index(op.f('ix_schedule_group'), 'schedule', ['group'], unique=False)
-    op.create_index(op.f('ix_schedule_id'), 'schedule', ['id'], unique=False)
-    op.create_index(op.f('ix_schedule_lesson_number'), 'schedule', ['lesson_number'], unique=False)
-    op.create_index(op.f('ix_schedule_module_id'), 'schedule', ['module_id'], unique=False)
-    op.create_index(op.f('ix_schedule_room'), 'schedule', ['room'], unique=False)
-    op.create_index(op.f('ix_schedule_semester'), 'schedule', ['semester'], unique=False)
-    op.create_index(op.f('ix_schedule_teacher_id'), 'schedule', ['teacher_id'], unique=False)
-    op.create_index(op.f('ix_schedule_weekday'), 'schedule', ['weekday'], unique=False)
-    op.create_table('users',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('username', sa.String(length=255), nullable=True),
-    sa.Column('email', sa.String(length=255), nullable=True),
-    sa.Column('hashed_password', sa.String(length=255), nullable=True),
-    sa.Column('is_active', sa.Boolean(), nullable=True),
-    sa.Column('student_id', sa.Integer(), nullable=True),
-    sa.Column('admin_id', sa.Integer(), nullable=True),
-    sa.Column('teacher_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['admin_id'], ['admins.id'], ),
-    sa.ForeignKeyConstraint(['student_id'], ['students.id'], ),
-    sa.ForeignKeyConstraint(['teacher_id'], ['teachers.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_users_admin_id'), 'users', ['admin_id'], unique=True)
-    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    op.create_index(op.f('ix_users_student_id'), 'users', ['student_id'], unique=True)
-    op.create_index(op.f('ix_users_teacher_id'), 'users', ['teacher_id'], unique=True)
-    op.create_index(op.f('ix_users_username'), 'users', ['username'], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_users_username'), table_name='users')
-    op.drop_index(op.f('ix_users_teacher_id'), table_name='users')
-    op.drop_index(op.f('ix_users_student_id'), table_name='users')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_index(op.f('ix_users_email'), table_name='users')
-    op.drop_index(op.f('ix_users_admin_id'), table_name='users')
-    op.drop_table('users')
-    op.drop_index(op.f('ix_schedule_weekday'), table_name='schedule')
-    op.drop_index(op.f('ix_schedule_teacher_id'), table_name='schedule')
-    op.drop_index(op.f('ix_schedule_semester'), table_name='schedule')
-    op.drop_index(op.f('ix_schedule_room'), table_name='schedule')
-    op.drop_index(op.f('ix_schedule_module_id'), table_name='schedule')
-    op.drop_index(op.f('ix_schedule_lesson_number'), table_name='schedule')
-    op.drop_index(op.f('ix_schedule_id'), table_name='schedule')
-    op.drop_index(op.f('ix_schedule_group'), table_name='schedule')
-    op.drop_index(op.f('ix_schedule_class_type'), table_name='schedule')
-    op.drop_table('schedule')
-    op.drop_table('weekdays_lessons_association')
-    op.drop_table('teachers_modules_association')
-    op.drop_index(op.f('ix_students_second_name'), table_name='students')
-    op.drop_index(op.f('ix_students_last_name'), table_name='students')
-    op.drop_index(op.f('ix_students_id'), table_name='students')
-    op.drop_index(op.f('ix_students_first_name'), table_name='students')
-    op.drop_index(op.f('ix_students_academic_year'), table_name='students')
-    op.drop_index(op.f('ix_students_academic_group'), table_name='students')
-    op.drop_table('students')
-    op.drop_index(op.f('ix_rooms_number'), table_name='rooms')
-    op.drop_index(op.f('ix_rooms_id'), table_name='rooms')
-    op.drop_index(op.f('ix_rooms_class_type_id'), table_name='rooms')
-    op.drop_table('rooms')
-    op.drop_table('modules_typesOfClasses_association')
-    op.drop_index(op.f('ix_years_number'), table_name='years')
-    op.drop_index(op.f('ix_years_id'), table_name='years')
-    op.drop_table('years')
-    op.drop_index(op.f('ix_weekdays_name'), table_name='weekdays')
-    op.drop_index(op.f('ix_weekdays_id'), table_name='weekdays')
-    op.drop_table('weekdays')
-    op.drop_index(op.f('ix_teachers_busy_weekday'), table_name='teachers_busy')
-    op.drop_index(op.f('ix_teachers_busy_teacher_id'), table_name='teachers_busy')
-    op.drop_index(op.f('ix_teachers_busy_lesson'), table_name='teachers_busy')
-    op.drop_index(op.f('ix_teachers_busy_is_busy'), table_name='teachers_busy')
-    op.drop_index(op.f('ix_teachers_busy_id'), table_name='teachers_busy')
-    op.drop_table('teachers_busy')
-    op.drop_index(op.f('ix_teachers_second_name'), table_name='teachers')
-    op.drop_index(op.f('ix_teachers_last_name'), table_name='teachers')
-    op.drop_index(op.f('ix_teachers_id'), table_name='teachers')
-    op.drop_index(op.f('ix_teachers_first_name'), table_name='teachers')
-    op.drop_table('teachers')
-    op.drop_index(op.f('ix_rooms_busy_weekday'), table_name='rooms_busy')
-    op.drop_index(op.f('ix_rooms_busy_room_id'), table_name='rooms_busy')
-    op.drop_index(op.f('ix_rooms_busy_lesson'), table_name='rooms_busy')
-    op.drop_index(op.f('ix_rooms_busy_is_busy'), table_name='rooms_busy')
-    op.drop_index(op.f('ix_rooms_busy_id'), table_name='rooms_busy')
-    op.drop_table('rooms_busy')
-    op.drop_index(op.f('ix_modules_year'), table_name='modules')
-    op.drop_index(op.f('ix_modules_name'), table_name='modules')
-    op.drop_index(op.f('ix_modules_id'), table_name='modules')
-    op.drop_table('modules')
-    op.drop_index(op.f('ix_lessons_time'), table_name='lessons')
-    op.drop_index(op.f('ix_lessons_number'), table_name='lessons')
-    op.drop_index(op.f('ix_lessons_id'), table_name='lessons')
-    op.drop_table('lessons')
-    op.drop_index(op.f('ix_groups_busy_weekday'), table_name='groups_busy')
-    op.drop_index(op.f('ix_groups_busy_lesson'), table_name='groups_busy')
-    op.drop_index(op.f('ix_groups_busy_is_busy'), table_name='groups_busy')
-    op.drop_index(op.f('ix_groups_busy_id'), table_name='groups_busy')
-    op.drop_index(op.f('ix_groups_busy_group_id'), table_name='groups_busy')
-    op.drop_table('groups_busy')
-    op.drop_index(op.f('ix_groups_number'), table_name='groups')
-    op.drop_index(op.f('ix_groups_id'), table_name='groups')
-    op.drop_table('groups')
-    op.drop_index(op.f('ix_class_types_name'), table_name='class_types')
-    op.drop_index(op.f('ix_class_types_id'), table_name='class_types')
-    op.drop_table('class_types')
-    op.drop_index(op.f('ix_admins_name'), table_name='admins')
-    op.drop_index(op.f('ix_admins_id'), table_name='admins')
-    op.drop_table('admins')
-    # ### end Alembic commands ###
