+++ test_app/db/__init__.py
@@ -1,0 +1,1 @@
+from database import *
