+++ test_app/enums/__init__.py
@@ -1,0 +1,1 @@
+from enum_models import *
